requirements.txt
# Core dependencies
openpyxl>=3.1.0
pandas>=2.0.0

# Configuration files
PyYAML>=6.0.0

# OpenAI API for GUID Formula Updater
openai>=1.0.0

# Optional AI agent dependencies
# Uncomment if using with LangChain/LangGraph
# langchain>=0.1.0
# langchain-community>=0.0.20
# langsmith>=0.0.80

# Development dependencies
# pytest>=7.0.0
# black>=23.0.0
# isort>=5.12.0
# mypy>=1.0.0